+++ sneakers.py
@@ -3,12 +3,11 @@ from selenium_stealth import stealth
 from selenium.webdriver.common.by import By
 from openpyxl import Workbook
 from selenium.webdriver.chrome.service import Service
-import added
 import sqlite3 as sq
 import openpyxl
 
 
-def export():
+def get_data():
     clear()
     wb = Workbook()
     ws = wb.active
@@ -19,7 +18,7 @@ def export():
     options.add_argument("start-maximized")
     options.add_experimental_option("excludeSwitches", ["enable-automation"])
     options.add_experimental_option('useAutomationExtension', False)
-    s = Service('C:/Users/denis/Downloads/chromedriver.exe')
+    s = Service('chromedriver.exe')
     driver = webdriver.Chrome(service=s)
     stealth(driver,
             languages=["en-US", "en"],
@@ -29,7 +28,7 @@ def export():
             renderer="Intel Iris OpenGL Engine",
             fix_hairline=True,
             )
-    for page in range(1, 3):
+    for page in range(1, 4):
         url = f"https://sneakerhead.ru/shoes/sneakers/?PAGEN_1={page}"
         driver.get(url)
         blocks = driver.find_element(By.CLASS_NAME, "product-cards__list")
@@ -50,7 +49,6 @@ def export():
                                                                                  "product-card__title").find_element(
                 By.CLASS_NAME, "product-card__link").get_attribute("href")
             ws.append([title, price, image, link])
-            print(title, price, image, link)
 
     wb.save('sneakers.xlsx')
     wb.close()
@@ -80,6 +78,13 @@ def export():
 def clear():
     con = sq.connect("sneakers.db")
     cur = con.cursor()
-    cur.execute("DELETE FROM sneakers")
-    con.commit()
-    con.close()
+    cur.execute("SELECT name FROM sqlite_master WHERE type='table';")
+    tables = cur.fetchall()
+    table_names = [table[0] for table in tables]
+    if 'sneakers' in table_names:
+        cur.execute("DELETE FROM sneakers")
+        con.commit()
+        con.close()
+    else:
+        con.commit()
+        con.close()
